Hamming distance
In information theory, the Hamming distance between two strings of equal
length is the number of positions at which the corresponding symbols are
different. In another way, it measures the minimum number of
substitutions required to change one string into the other, or the
minimum number of errors that could have transformed one string into the
other.